Update testing suite
